+++ SampleProjects/TestSomething/test/godmode.cpp
@@ -34,17 +34,6 @@ unittest(random)
   assertEqual(state->seed, 4294967282);
 }
 
-void myInterruptHandler() {
-}
-
-unittest(interrupts)
-{
-  // these are meaningless for testing; just call the routine directly.
-  // make sure our mocks work though
-  attachInterrupt(2, myInterruptHandler, CHANGE);
-  detachInterrupt(2);
-}
-
 unittest(pins)
 {
   GodmodeState* state = GODMODE();
@@ -170,17 +159,6 @@ unittest(pin_write_history)
 
   assertEqual("No", state->digitalPin[3].toAscii(1, false));
 
-}
-
-unittest(interrupt_attachment) {
-  GodmodeState *state = GODMODE();
-  state->reset();
-  assertFalse(state->interrupt[0].attached);
-  attachInterrupt(0, (void (*)(void))0, 3);
-  assertTrue(state->interrupt[0].attached);
-  assertEqual(state->interrupt[0].mode, 3);
-  detachInterrupt(0);
-  assertFalse(state->interrupt[0].attached);
 }
 
 unittest(spi) {